+++ src/api/company/application/cqrs/query/get-company-with-profile-id.query.ts
@@ -1,0 +1,7 @@
+export class GetCompanyWithProfileIdQuery {
+  public readonly profileId: string;
+
+  constructor(partial: Partial<GetCompanyWithProfileIdQuery>) {
+    Object.assign(this, partial);
+  }
+}