+++ src/api/company/application/cqrs/query/get-company-with-user-id.query.ts
@@ -1,0 +1,7 @@
+export class GetCompanyWithUserIdQuery {
+  public readonly userId: string;
+
+  constructor(partial: Partial<GetCompanyWithUserIdQuery>) {
+    Object.assign(this, partial);
+  }
+}