+++ src/api/company/application/cqrs/query/get-employee-by-company-id.query.ts
@@ -1,0 +1,7 @@
+export class GetEmployeeByCompanyIdQuery {
+  public readonly companyId: string;
+
+  constructor(partial: Partial<GetEmployeeByCompanyIdQuery>) {
+    Object.assign(this, partial);
+  }
+}