+++ src/api/company/application/cqrs/query/is-role-in-company.query.ts
@@ -1,11 +1,11 @@
-import {RoleCompanyEmployeeEnum} from "../../../domain/enum/role-company-employee.enum";
+import { RoleCompanyEmployeeEnum } from '../../../domain/enum/role-company-employee.enum';
 
 export class IsRoleInCompanyQuery {
-  constructor(partial: Partial<IsRoleInCompanyQuery>) {
-    Object.assign(this, partial);
-  }
-
   public readonly companyId: string;
   public readonly userId: string;
   public readonly roles: RoleCompanyEmployeeEnum[];
+
+  constructor(partial: Partial<IsRoleInCompanyQuery>) {
+    Object.assign(this, partial);
+  }
 }