+++ src/api/company/web/request/get-company-discovery.request.ts
@@ -1,0 +1,10 @@
+export class GetCompanyDiscoveryRequest {
+  public readonly userX?: string;
+  public readonly userY?: string;
+  public readonly distance?: number;
+  public readonly occupationIdList: string[];
+
+  constructor(partial: Partial<GetCompanyDiscoveryRequest>) {
+    Object.assign(this, partial);
+  }
+}