+++ src/api/company/web/request/remove-company-employee.request.ts
@@ -1,0 +1,8 @@
+export class RemoveCompanyEmployeeRequest {
+  public readonly companyId: string;
+  public readonly profileId: string;
+
+  constructor(partial: Partial<RemoveCompanyEmployeeRequest>) {
+    Object.assign(this, partial);
+  }
+}