+++ src/api/profile/domain/entities/profile.entity.ts
@@ -13,10 +13,10 @@ import {
 } from 'typeorm';
 import { UserEntity } from '../../../user/domain/entities/user.entity';
 import { OccupationEntity } from '../../../occupation/domain/entities/occupation.entity';
-import { CompanyEmployeeEntity } from '../../../entities-to-create/company-employee.entity';
+import { CompanyEmployeeEntity } from '../../../company/domain/entities/company-employee.entity';
 import { MediaEntity } from '../../../entities-to-create/media.entity';
 import { JoinedConversation } from '../../../entities-to-create/joined-conversation.entity';
-import CompanyEntity from '../../../entities-to-create/company.entity';
+import CompanyEntity from '../../../company/domain/entities/company.entity';
 import { NotificationEntity } from '../../../entities-to-create/notification.entity';
 import { RoleProfileEnum } from '../enum/role-profile.enum';
 import { CardEntity } from '../../../card/domain/entities/card.entity';