+++ src/app.module.ts
@@ -24,6 +24,7 @@ import { CardModule } from './api/card/card.module';
 import { GroupModule } from './api/groupe/group.module';
 import * as process from 'process';
 import { AddressModule } from './api/address/address.module';
+import { CompanyModule } from './api/company/company.module';
 
 @Module({
   imports: [
@@ -53,6 +54,7 @@ import { AddressModule } from './api/address/address.module';
     CardModule,
     GroupModule,
     AddressModule,
+    CompanyModule,
 
     // ________ Module to remove ________
     EntitiesToMoveModule,